BUG/MINOR; Ensured that subsequent auto update operations can occur
If there is a problem with the temporary file delete operations or the client can not be disposed of the auto download signal is still set to enable other auto update operations to continue.

+++ FoundationV3/Mobile/Detection/AutoUpdate.cs
@@ -255,10 +255,25 @@ namespace FiftyOne.Foundation.Mobile.Detection
             }
             finally
             {
-                File.Delete(compressedTempFile);
-                File.Delete(uncompressedTempFile);
-                client.Dispose();
-                _autoDownloadUpdateSignal.Set();
+                try
+                {
+                    // Delete the temporary files if they exit.
+                    if (File.Exists(compressedTempFile))
+                    {
+                        File.Delete(compressedTempFile);
+                    }
+                    if (File.Exists(uncompressedTempFile))
+                    {
+                        File.Delete(uncompressedTempFile);
+                    }
+                    client.Dispose();
+                }
+                finally
+                {
+                    // Ensure that whatever happens the signal is reset to
+                    // allow another auto update operation to proceed.
+                    _autoDownloadUpdateSignal.Set();
+                }
             }
             return status;
         }
